% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAccXGboost.R
\name{getAccXGboost}
\alias{getAccXGboost}
\title{Estimate the accuracy of the trained xgboost model}
\usage{
getAccXGboost(model, data, outcome)
}
\arguments{
\item{model}{output xgboost model from `buildXGboost()`}

\item{data}{dataframe, input data used to train the model}

\item{outcome}{string, name of the outcome variable}
}
\value{
a number from 0 to 1 indicating the accuracy of the model (1 being perfect accuracy)
}
\description{
Estimate the accuracy of the trained xgboost model
}
\examples{
\dontrun{
getAccXGboost(xgb, voteData, "vote")
}
}
