% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeatureImp.R
\name{getFeatureImp}
\alias{getFeatureImp}
\title{Calculate feature importance in terms of xgboost metrics and shap values}
\usage{
getFeatureImp(model, data, outcome)
}
\arguments{
\item{model}{output xgboost model from `buildXGboost()`}

\item{data}{dataframe, input data used to train the model}

\item{outcome}{string, name of the outcome variable}
}
\value{
a data frame that contains 4 metrics for each feature
\itemize{
 \item Gain: improvement in accuracy brought by the feature to the branches it is on
 \item Cover: relative quantity of observations concerned by a feature
 \item Frequency: the number of times a feature is used in all generated trees
 \item shap: average Shapley value for the feature
}
}
\description{
Calculate feature importance in terms of xgboost metrics and shap values
}
\examples{
\dontrun{
getFeatureImp(xgb, voteData, "vote")
}
}
