% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getQCAMetric.R
\name{getQCAMetric}
\alias{getQCAMetric}
\title{Calculate per-variable QCA metric}
\usage{
getQCAMetric(dataCali, outcome)
}
\arguments{
\item{dataCali}{dataframe, calibrated data}

\item{outcome}{string, name of the outcome variable}
}
\value{
a data frame that contains 3 QCA metrics
\itemize{
 \item Cons.Nec: necessity consistency
 \item Cov.Nec: necessity coverage
 \item RoN: relavance of necessity
}
}
\description{
Calculate per-variable QCA metric
}
\examples{
\dontrun{
getQCAMetric(voteDataCalibrated, "vote")
}

}
