% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getXGboostCut.R
\name{getXGboostCut}
\alias{getXGboostCut}
\title{Extract the cutoffs used in the trained xgboost model}
\usage{
getXGboostCut(model, full = F, treeIndex = NULL)
}
\arguments{
\item{model}{output xgboost model from `buildXGboost()`}

\item{full}{logical, whether to report the most frequent cutoff for each feautres or all the cutoffs (default: False)}

\item{treeIndex}{number, examine cutoffs from a particular tree}
}
\value{
dataframe that contains the cutoffs used by the xgboost model for each feature
}
\description{
Extract the cutoffs used in the trained xgboost model
}
\examples{
\dontrun{
getXGboostCut(xgb)
}
}
