% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinMetrics.R
\name{joinMetrics}
\alias{joinMetrics}
\title{Helper function to join the feature-level QCA and XGboost metrics}
\usage{
joinMetrics(featureMetric, qcaMetric)
}
\arguments{
\item{featureMetric}{dataframe, result from the `getQCAMetric()`}

\item{qcaMetric}{dataframe, result from the `getFeatureImp()`}
}
\value{
dataframe, combined dataframe of feature-level importance
}
\description{
Helper function to join the feature-level QCA and XGboost metrics
}
\examples{
\dontrun{
fDf <- getFeatureImp(xgb, voteData, "vote")
qDf <- getQCAMetric(voteDataCalibrated, "vote")
joinDf <- joinMetrics(fDf, qDf)
}
}
