% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reCalibrateXGCut.R
\name{reCalibrateXGCut}
\alias{reCalibrateXGCut}
\title{Calibrate data based on xgboost model cutoffs}
\usage{
reCalibrateXGCut(data, cutoffs, outcome, na.rm = T)
}
\arguments{
\item{data}{the input data that was used to train xgboost model}

\item{cutoffs}{dataframe, contains cutoffs extracted from trained xgboost model}

\item{outcome}{string, indicates the name of the outcome variable}

\item{na.rm}{logical, whether to remove rows that contain NA after calibration (default: True)}
}
\value{
a calibrated dataframe
}
\description{
this function performs crisp calibration based on a set of cutoffs derived from the xgboost trees
}
\examples{
\dontrun{
reCalibrateXGCut(voteData, cutDataFrame, "vote")
}
}
