% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetSuffiTest.R
\name{subsetSuffiTest}
\alias{subsetSuffiTest}
\title{Perform sufficiency test on a select set of variables}
\usage{
subsetSuffiTest(
  dataCali,
  vSelect,
  outcome,
  incl.cut = 0.8,
  pri.cut = 0.51,
  n.cut = 10,
  sort.by = c("OUT", "incl"),
  complete = T
)
}
\arguments{
\item{dataCali}{dataframe, calibrated data}

\item{vSelect}{string vector, selected variable names (must be in dataCali)}

\item{outcome}{string, name of the outcome variable}

\item{incl.cut}{inclusion cut-off(s); see `?QCA::truthTable` for details}

\item{pri.cut}{minimal score for the PRI; see `?QCA::truthTable` for details}

\item{n.cut}{minimum number of cases for a remainder row; see `?QCA::truthTable` for details}

\item{sort.by}{Sort the truth table according to various columns}

\item{complete}{Logical, whether to print complete truth table}
}
\value{
a list of two tables
\itemize{
 \item table: truth table
 \item path: sufficiency test solution
}
}
\description{
Perform sufficiency test on a select set of variables
}
\examples{
\dontrun{
subsetSuffiTest(voteDataCalibrated, c("P_V530", "V648", "V207", "V141"), "vote")
}
}
