% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumSuffTestRes.R
\name{sumSuffTestRes}
\alias{sumSuffTestRes}
\title{helper function to combine results from a series of sufficiency tests}
\usage{
sumSuffTestRes(featureList, resList)
}
\arguments{
\item{featureList}{a list of feature vectors}

\item{resList}{a list of corresponding results from `subsetSuffiTest()` on the above featureList}
}
\value{
dataframe, that summarizes the results from the sufficiency tests
}
\description{
helper function to combine results from a series of sufficiency tests
}
