% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voteData.R
\docType{data}
\name{voteData}
\alias{voteData}
\title{panel data of youth voting}
\format{
## `voteData`
A data frame with 427 rows and 77 columns:
\describe{
  \item{vote}{whether the person voted or not}
  ...
}
}
\source{
<https://www.who.int/teams/global-tuberculosis-programme/data>
}
\usage{
voteData
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
